Applied requested changes, improved application test

+++ src/main/java/org/prebid/server/proto/openrtb/ext/request/somoaudience/ExtImpSomoaudience.java
@@ -3,11 +3,13 @@ package org.prebid.server.proto.openrtb.ext.request.somoaudience;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor(staticName = "of")
 @Value
 public class ExtImpSomoaudience {
 
     String placementHash;
 
-    Float bidFloor;
+    BigDecimal bidFloor;
 }
